refactor(api): PR-6 commit 3 — installs router (10 routes 5 子域合一)
server.py 3786 → 3548 行（−238），10 install / runtime / LLM tagger admin
routes + 5 inline BaseModel 抽出。

  api/routers/installs.py    10 routes  wd14 (2) + torch (2) + flash-attn (2)
                                        + xformers (2) + llm-tagger admin (2)
  api/schemas/installs.py    5 BaseModel  WD14Install / TorchReinstall /
                                          FlashAttnInstall / LLMModelsRefresh /
                                          LLMConnectionTest

5 个子域合一个 router 的判断：
- 每个子域只 2 route，单独 router 太碎
- 全部共用 install / restart_required 模式 + Settings 抽屉 UI 装在同一处
- _select_preset llm-tagger 私有 helper 留在 router 内

回归：PR-1 安全网 152 + studio_server 44 + llm_tagger 2 + ... = **242/242
全绿**。route snapshot 不变（160 routes）。
